✨ refactor: left-panel state is now saved in a cookie
