+++ apps/web/app/providers.tsx
@@ -1,14 +1,15 @@
 import { ReactNode } from 'react'
-import { AppShellContextProvider } from '@/components/app-shell/app-shell-context'
+import { AppShellContextProvider } from '@/components/app-shell/app-shell-provider'
 import { TooltipProvider } from '@ror/react/components/tooltip'
 
 interface ProvidersProps {
   children: ReactNode
+  defaultSidebarOpen: boolean
 }
 
-export function Providers({ children }: ProvidersProps) {
+export function Providers({ children, defaultSidebarOpen }: ProvidersProps) {
   return (
-    <AppShellContextProvider>
+    <AppShellContextProvider defaultSidebarOpen={defaultSidebarOpen}>
       <TooltipProvider>{children}</TooltipProvider>
     </AppShellContextProvider>
   )
